ssiog/requirements.txt
gcsfs==2024.6.1
pyarrow==17.0.0
torch==2.4.0
grpcio==1.67.1
opentelemetry-exporter-gcp-monitoring==1.7.0a0
opentelemetry-resourcedetector-gcp==1.7.0a0
opentelemetry-propagator-gcp==1.7.0
psutil==6.1.1
pandas==2.2.3
tqdm==4.67.1
pytest==8.3.4
